recruitment-mcp/requirements.txt
fastapi
uvicorn
python-dotenv
requests
mcp
httpx